pyproject.toml
[tool.poetry]
name = "worldticket-backend"
version = "0.1.0"
description = "World Ticket Backend API — FIFA World Cup 2026 Ticketing Platform"
authors = ["World Ticket Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.6.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.12"
cloudinary = "^1.41.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
celery = {extras = ["redis"], version = "^5.4.0"}
stripe = "^11.3.0"
slowapi = "^0.1.9"
httpx = "^0.28.0"
python-dotenv = "^1.0.1"
greenlet = "^3.1.1"
gevent = "^26.4.0"
bcrypt = "4.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
httpx = "^0.28.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
